refactor(demo): improve file structure
This commit improves the file structure for the application contexts and
hooks.
